change num int to double

+++ src/main/java/com/thoughtworks/iamcoach/pos/CartItem.java
@@ -16,7 +16,7 @@ public class CartItem {
         return num;
     }
 
-    public void setNum(int num) {
+    public void setNum(double num) {
         this.num = num;
     }
